Refactor(server): MySQL Pool을 불러오는 로직 분리
- 해당 로직을 server/config/db.js 파일로 이동
- module.exports를 사용하여 index.js 파일에서 pool을 불러오도록 수정

ref: #39
