+++ packages/server/index.js
@@ -1,6 +1,7 @@
 const express = require('express');
-const mysql = require('mysql2');
 const cors = require('cors');
+const db = require('./config/db');
+
 require('dotenv').config();
 
 const app = express();
@@ -10,23 +11,6 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
 app.use(cors());
-
-const dbConfig = {
-  host: process.env.HOST,
-  user: process.env.USERNAME,
-  password: process.env.PASSWORD,
-  port: process.env.PORT,
-};
-
-const pool = mysql.createPool(dbConfig);
-
-pool.getConnection(err => {
-  if (err) {
-    console.error('MySQL 연결 실패:', err);
-  } else {
-    console.log('MySQL 연결 성공');
-  }
-});
 
 app.get('/', (req, res) => {
   console.log('get /');
